Disable mouseover on the encounter mechanic frames in the list
We'd love to have a good tooltip here, but until we have a good tooltip,
these frames should ignore the mouse entirely. That prevents the
annoying flashing of the row highlight as you mouse over a row of
mechanics.

Also tweak the template we use slightly. We don't need the checkmark
support, so we can use one of the parent templates instead. It doesn't
really affect anything, it just seemed unnecessary to pull in the
checkmark machinery if we weren't going to use it.